Ensure we can populate the model config with model values
Setting the model config fraught with danger. I'm probably not doing
it 100% right. At any fashion this will get us up and running and
Tom can assess if this is correct.

I've had to add better error handling, to aid with debugging, but
it's required. Also the model defaults need tests, as there is
a bug in the select statements which I had to fix.
